# app/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIDI Agent API",
    "description": "AI-powered MIDI generation from natural language prompts",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/midi/tracks/": {
      "post": {
        "tags": ["tracks"],
        "summary": "Create Track",
        "description": "Create a new MIDI track for a song.\n\nThe MIDI channel is automatically assigned as the next available channel\nbased on existing tracks in the song (max channel + 1).\n\nThe track color is automatically assigned cyclically from the color palette.",
        "operationId": "create_track_api_midi_tracks__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/midi/tracks/{track_id}": {
      "delete": {
        "tags": ["tracks"],
        "summary": "Delete Track",
        "description": "Delete a track and all associated loops.\n\nThe loops are automatically deleted via cascade relationship.",
        "operationId": "delete_track_api_midi_tracks__track_id__delete",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["tracks"],
        "summary": "Update Track",
        "description": "Update a track's title, MIDI channel, instrument, and/or color.\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_track_api_midi_tracks__track_id__patch",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTrackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTrackRequest": {
        "properties": {
          "song_id": {
            "type": "string",
            "title": "Song Id",
            "description": "ID of the parent song"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Track title"
          }
        },
        "type": "object",
        "required": ["song_id", "title"],
        "title": "CreateTrackRequest",
        "description": "Request model for creating a new track."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoopResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Loop ID"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Position offset in measures from start of song"
          },
          "measures": {
            "type": "integer",
            "title": "Measures",
            "description": "Number of measures in the loop"
          },
          "extend_measures": {
            "type": "integer",
            "title": "Extend Measures",
            "description": "Number of extra measures to add (or subtract) from the loop"
          },
          "midi_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Midi Events",
            "description": "MIDI events in the loop"
          },
          "track_id": {
            "type": "string",
            "title": "Track Id",
            "description": "ID of the parent track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          }
        },
        "type": "object",
        "required": [
          "id",
          "offset",
          "measures",
          "extend_measures",
          "midi_events",
          "track_id",
          "created_at",
          "updated_at"
        ],
        "title": "LoopResponse",
        "description": "Response model for MIDI loops, excluding chat message history."
      },
...
      "PatchTrackRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Track title (optional)"
          },
          "midi_channel": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 16.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Midi Channel",
            "description": "MIDI channel 1-16 (optional)"
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["piano", "bass", "drum"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument",
            "description": "Instrument type (optional)"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["error", "warning", "success", "info", "primary", "secondary"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Color",
            "description": "Color theme (optional)"
          }
        },
        "type": "object",
        "title": "PatchTrackRequest",
        "description": "Request model for updating a track."
      },
...
      "TrackDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Track ID"
          },
          "song_id": {
            "type": "string",
            "title": "Song Id",
            "description": "ID of the parent song"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Track title"
          },
          "midi_channel": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Midi Channel",
            "description": "MIDI channel (1-16)"
          },
          "instrument": {
            "type": "string",
            "enum": ["piano", "bass", "drum"],
            "title": "Instrument",
            "description": "Instrument type for this track"
          },
          "color": {
            "type": "string",
            "enum": ["error", "warning", "success", "info", "primary", "secondary"],
            "title": "Color",
            "description": "Color theme for this track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          },
          "loops": {
            "items": {
              "$ref": "#/components/schemas/LoopResponse"
            },
            "type": "array",
            "title": "Loops",
            "description": "Loops in this track"
          }
        },
        "type": "object",
        "required": [
          "id",
          "song_id",
          "title",
          "midi_channel",
          "instrument",
          "color",
          "created_at",
          "updated_at",
          "loops"
        ],
        "title": "TrackDetailResponse",
        "description": "Response model for MIDI tracks including loops."
      },
      "TrackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Track ID"
          },
          "song_id": {
            "type": "string",
            "title": "Song Id",
            "description": "ID of the parent song"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Track title"
          },
          "midi_channel": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Midi Channel",
            "description": "MIDI channel (1-16)"
          },
          "instrument": {
            "type": "string",
            "enum": ["piano", "bass", "drum"],
            "title": "Instrument",
            "description": "Instrument type for this track"
          },
          "color": {
            "type": "string",
            "enum": ["error", "warning", "success", "info", "primary", "secondary"],
            "title": "Color",
            "description": "Color theme for this track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          }
        },
        "type": "object",
        "required": ["id", "song_id", "title", "midi_channel", "instrument", "color", "created_at", "updated_at"],
        "title": "TrackResponse",
        "description": "Response model for MIDI tracks (excludes loops)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}